refactor(core/modules): subdirectory structure -> avoids long repetitive names
