+++ apps/core/src/index.ts
@@ -12,8 +12,8 @@ import apiRoutes from './routes/api';
 import internalRoutes from './routes/internal';
 import { ProcessManager } from './modules/process.manager';
 import { GameManager } from './modules/game.manager';
-import { ConfigManager } from './modules/config/config.manager';
-import { perfManager } from './modules/perf/perf.manager';
+import { ConfigManager } from './modules/config/manager';
+import { perfManager } from './modules/perf/manager';
 import { applyMigrations } from '@fxmanager/database';
 
 applyMigrations();
