+++ apps/core/src/middleware/resource.ts
@@ -1,5 +1,5 @@
 import type { FastifyRequest, FastifyReply } from 'fastify';
-import { ConfigManager } from '../modules/config/config.manager';
+import { ConfigManager } from '../modules/config/manager';
 
 export async function resourceAuth(req: FastifyRequest, reply: FastifyReply) {
 	const token = req.headers['x-resource-token'];
