+++ apps/core/src/modules/config/manager.ts
