+++ apps/core/src/modules/perf/manager.ts
@@ -3,8 +3,8 @@ import {
 	type PerfSnapshot,
 	type RawPerf,
 } from '@fxmanager/shared/types';
-import { diffPerfs, didPerfReset, parseRawPerf } from './perf/perf.parser';
-import { wsManager } from './ws.manager';
+import { diffPerfs, didPerfReset, parseRawPerf } from './parser';
+import { wsManager } from '../ws.manager';
 
 // TODO: replace this with a utility function for getting the actual server endpoint at a later date
 const PERF_PORT = 30120;
