+++ apps/core/src/modules/perf/parser.ts
